requirements.txt
# =====================================================================
# REQUIREMENTS.TXT - SIGETI DWH PROJECT
# Python Data Warehouse with dbt + Prefect + PostgreSQL
# Generated: 2025-11-26
# Python: 3.12
# =====================================================================

# ============= CORE DBT =============
dbt-core==1.10.15
dbt-postgres==1.9.1
dbt-adapters==1.20.1
dbt-common==1.36.0
dbt-extractor==0.6.0
dbt-semantic-interfaces==0.9.0
dbt-protos==1.0.397

# ============= DATA PROCESSING =============
pandas==2.1.4
numpy==1.26.4
agate==1.9.1
tqdm==4.67.1

# ============= DATABASE & ORM =============
psycopg2-binary==2.9.9
SQLAlchemy>=2.0.0
alembic==1.17.1
aiosqlite==0.21.0

# ============= VALIDATION & SERIALIZATION =============
pydantic==2.12.4
pydantic_core==2.41.5
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
protobuf==6.33.1
orjson==3.11.4

# ============= WEB FRAMEWORK & API =============
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-ratelimit==4.1.0

# ============= AUTHENTICATION & SECURITY =============
PyJWT==2.10.1
cryptography==46.0.3
certifi==2025.11.12
oauthlib==3.3.1

# ============= HTTP & REQUESTS =============
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9

# ============= DATE & TIME =============
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
parsedatetime==2.6
pytimeparse==1.1.8

# ============= UTILITIES =============
python-dotenv==1.0.0
click==8.3.1
python-slugify==8.0.4
text-unidecode==1.3
packaging==25.0
typing_extensions==4.15.0
attrs==25.4.0
more-itertools==10.8.0
six==1.17.0
deepdiff==8.6.1
orderly-set==5.5.0

# ============= TEMPLATES =============
Jinja2==3.1.6
MarkupSafe==3.0.3
PyYAML==6.0.3

# ============= NETWORKING & ENCODING =============
urllib3==2.5.0
charset-normalizer==3.4.4
idna==3.11
h11==0.16.0
anyio==4.11.0
sniffio==1.3.1

# ============= MONITORING & LOGGING =============
openai==1.12.0
python-json-logger==4.0.0

# ============= SYSTEM & BUILD =============
mashumaro==3.14
pathspec==0.12.1
importlib_metadata==8.7.0
isodate==0.6.1
distro==1.9.0
asgiref==3.10.0
babel==2.17.0
daff==1.4.2
leather==0.4.0
referencing==0.37.0
rpds-py==0.29.0
snowplow-tracker==1.1.0
sqlparse==0.5.3
typing-inspection==0.4.2
annotated-types==0.7.0
colorama==0.4.6
networkx==3.6
